Tutorial: Add background music
This tutorial will get you started with audio by introducing you to the Audio Source Component and showing you how audio is added to a Scene.
In this tutorial, you will:
Add an Audio Source Component to a GameObject.
Add an Audio Clip to an Audio Source.
Implement background music in a 3D Scene.