Tutorial: Create real-time 3D audio effects
Sound in real-time 3D must behave like sound does in the real world: it is louder as you get closer to its source, and even bounces off objects. In this tutorial, you will get a first look at the Unity features that make audio come to life.
In this tutorial, you will be able to:
Define 3D sound.
Explain the role of an Audio Listener in a Scene.
Set an Audio Clip to loop.
Convert background audio into 3D sound.